fix single/double outer quote swap mechanism, move to common.py
